fix(Market): Fixed that you get the marketFee now correctly. My mistake was that the variablesMap had to be a String Integer Map and not String Double as the MarketFee is a double and not integer.

+++ src/main/java/de/uniks/stp24/rest/PresetsApiService.java
@@ -24,7 +24,7 @@ public interface PresetsApiService {
     Observable<ArrayList<DistrictPresets>> getDistrictPresets();
 
     @GET("presets/variables")
-    Observable<Map<String, Integer>> getVariables();
+    Observable<Map<String, Double>> getVariables();
 
     @GET("presets/resources")
     Observable<ResourceDto> getResources();
